+++ django_datajsonar/admin/__init__.py
@@ -1,0 +1,10 @@
+from django_datajsonar.models import Jurisdiction, Stage
+from .synchronizer import *
+from .repeatable_job import *
+from .data_json import *
+from .metadata import *
+from .node import *
+from .tasks import *
+
+admin.site.register(Jurisdiction)
+admin.site.register(Stage)
